Initial implementation of catch previews.
Uses the `ReplayableRandom` (modified to be per-screen) and many transpilers to force `Game1.random` to be redirected to the `ReplayableRandom` for fishing only.

Adds configuration settings (no GMCM yet) for both splash and catch previews.

Freeze operation is supported for catches but not yet attached to rod casts.

Includes minor bugfix for legendary fish in frenzy predictions.