Fix settings menu initialization when starting with custom rules.
Previously the logic was to load profile rules as selected rules, different from the default-initialized custom rules. However, if the profile rules are already custom then we instead need to override the default custom rules with those from the profile and make sure the "selected" and "custom" references are the same.