Ensure catch previews get updated after a freeze.
As previously implemented, freezing would allow the RNG update to be postponed indefinitely, by repeatedly casting and canceling near the respawn time. While this probably isn't a mechanic that can be usefully abused, it wasn't really intended either.

A second issue is that the time-change logic was simply broken and only work for the (formerly) default 10-minute respawn interval, since it only compared the last frame's time of day to the current.

This fixes both issues by tracking when the last actual snapshot was, and adds an extra non-forcing check to fishing cancellation to ensure overdue updates are handled immediately.

Changes the default interval to 30 minutes since the rest is now easily customizable.