Mega update/fix for timescale hack.
First major fix is to switch the outright mutation of NPC speed to a Harmony postfix, which fortunately does not seem to be mooted by inlining. Since the `speed` is set in totally arbitrary places to totally arbitrary values, we really cannot rely on "dividing" the speed to undo the timescale hack, and sometimes it will result in NPCs getting stuck as their speed gets set to 1 or even 0. The Harmony patch might not be universally perfect (see comments) but should be good enough.

Second major fix is for NPC pathfinding, which is extremely fragile to begin with. The solution used here, which has been successful over several tests, is to detect and correct "overshoot", i.e. when a controlled character crosses from one edge of a target bounding box to the other edge without ever landing in the "middle" where it gets detected as having reached that node. When this happens, the patch moves the NPC back to the closest spot that _is_ inside the tile's bounding box.

Minor fixes include applying the speed multiplier to various delays as well as the speed, e.g. when NPCs run into obstacles.