Add GMCM "integration".
This is primarily a stub menu, with some placeholder text in case of failure, and a Harmony patch to open the custom menu instead (and handle transition back to GMCM).

Removes the hardcoded F7 keybind and makes it a (hidden) setting. Main reason to keep it around is simply to provide an alternate means of accessing the menu if the GMCM patching fails.