Fix splash visibility condition.
This only came up when transitioning areas during a splash. The condition was set to find the first splash starting on or after the current game time, but that will skip to the "next" splash if one is already started. Correct logic is to find the first splash **ending after** the current time, which will include those already started but not finished.